pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["macalistervadim <macalistervadim@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
gunicorn = "^23.0.0"
djangorestframework-simplejwt = "^5.4.0"
django-cleanup = "^9.0.0"
drf-spectacular = "^0.28.0"
drf-spectacular-sidecar = "^2025.3.1"
django-debug-toolbar = "^5.0.1"
channels = "^4.2.0"
daphne = "^4.1.2"
channels-redis = "^4.2.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.3"
mypy = "^1.15.0"
pre-commit = "^4.1.0"
django-stubs = "^5.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    "*/migrations/*",
    "venv/",
    "__init__.py",
    "apps.py",
]
line-length = 79
indent-width = 4

[tool.ruff.lint]
select = ["E", "W", "C", "F", "N", "COM", "I", "Q", "U", "N", "A", "M"]
ignore = ["F405"]
fixable = ["ALL"]
unfixable = []